+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [*Unreleased*]
+
+### Changed
+
+* `onQueryLink` is no longer filtering on `commandId` (as that is not sent in the `composeExtension/onQueryLink`)
+* Changed signature for `onQueryLink` to use `IAppBasedLinkQuery` as value, to match official swagger 
+
 ## [*1.3.0*] - <*2019-05-22*>
 
 ### Changed